Refactor stub building into standalone class.
