+++ sbncode/TPCReco/VertexStub/CMakeLists.txt
@@ -1,3 +1,31 @@
+art_make_library(LIBRARY_NAME sbncode_StubBuilder
+        SOURCE StubBuilder.cxx
+        LIBRARIES
+	${ART_FRAMEWORK_CORE}
+	${ART_FRAMEWORK_SERVICES_REGISTRY}
+	${ART_ROOT_IO_TFILESERVICE_SERVICE}                ${ART_FRAMEWORK_SERVICES}
+	${ART_FRAMEWORK_PRINCIPAL}
+	art_Persistency_Common
+	art_Persistency_Provenance
+	art_Utilities canvas
+	${MF_MESSAGELOGGER}
+	${MF_UTILITIES}
+	${FHICLCPP}
+	cetlib cetlib_except
+	${ROOT_EVE_LIB_LIST}
+	${ROOT_X3d}
+	${ROOT_BASIC_LIB_LIST}
+	MF_MessageLogger
+
+	sbnobj_Common_Reco
+	lardataobj_AnalysisBase
+        larcorealg_Geometry
+        larcore_Geometry_Geometry_service
+	larreco_Calorimetry
+	lardataobj_RecoBase
+	lardataalg_DetectorInfo
+)
+
 simple_plugin(VertexChargeVacuum module
 	${ART_FRAMEWORK_CORE}
 	${ART_FRAMEWORK_SERVICES_REGISTRY}
@@ -45,6 +73,7 @@ simple_plugin(VertexStubTracker module
         larcore_Geometry_Geometry_service
 	larreco_Calorimetry
 	lardataalg_DetectorInfo
+	sbncode_StubBuilder
 )
 
 install_headers()
